requirements.txt
torch==1.13
torchvision
opencv-python
tqdm
transformer
pytorch_lightning
pycocotools
scipy
albumentations